+++ lib/src/datastore/mutations.dart
@@ -15,18 +15,20 @@ class MutationBatch {
   /// The given entity will be added to the datastore if an entity with the
   /// key does not exist yet or the entity has an incomplete key (in which
   /// case a unique id is generated).
-  void insert(Entity e) => mutations.add(
-        new ds.Mutation()..insert = e.toApiObject(),
-      );
+  MutationBatch insert(Entity e) {
+    mutations.add(new ds.Mutation()..insert = e.toApiObject());
+    return this;
+  }
 
   /// Adds an `update` mutation to this batch.
   ///
   /// The given entity will overwritten in the datastore, an entity with the
   /// key must exist for this mutation to succeed. The key of the entity must
   /// be complete.
-  void update(Entity e) => mutations.add(
-        new ds.Mutation()..update = e.toApiObject(),
-      );
+  MutationBatch update(Entity e) {
+    mutations.add(new ds.Mutation()..update = e.toApiObject());
+    return this;
+  }
 
   /// Adds an `upsert` mutation to this batch.
   ///
@@ -34,35 +36,39 @@ class MutationBatch {
   /// is complete this is done regardless of whether an entity with the same key
   /// exists. The key can also be incomplete in which case a unique id will be
   /// generated and an insertion will be performed.
-  void upsert(Entity e) => mutations.add(
-        new ds.Mutation()..upsert = e.toApiObject(),
-      );
+  MutationBatch upsert(Entity e) {
+    mutations.add(new ds.Mutation()..upsert = e.toApiObject());
+    return this;
+  }
 
   /// Adds a `delete` mutation to this batch.
   ///
   /// Deletes the entity with the key. The key must be complete. The entity with
   /// the key may or may not exist.
-  void delete(Key key) => mutations.add(
-        new ds.Mutation()..delete = key.toApiObject(),
-      );
+  MutationBatch delete(Key key) {
+    mutations.add(new ds.Mutation()..delete = key.toApiObject());
+    return this;
+  }
 
   /// Adds multiple `insert` mutations to this batch.
   ///
   /// The given entity will be added to the datastore if an entity with the
   /// key does not exist yet or the entity has an incomplete key (in which
   /// case a unique id is generated).
-  void insertAll(Iterable<Entity> entities) => mutations.addAll(
-        entities.map((e) => new ds.Mutation()..insert = e.toApiObject()),
-      );
+  MutationBatch insertAll(Iterable<Entity> entities) {
+    mutations.addAll(entities.map((e) => new ds.Mutation()..insert = e.toApiObject()));
+    return this;
+  }
 
   /// Adds multiple `update` mutations to this batch.
   ///
   /// The given entity will overwritten in the datastore, an entity with the
   /// key must exist for this mutation to succeed. The key of the entity must
   /// be complete.
-  void updateAll(Iterable<Entity> entities) => mutations.addAll(
-        entities.map((e) => new ds.Mutation()..update = e.toApiObject()),
-      );
+  MutationBatch updateAll(Iterable<Entity> entities) {
+    mutations.addAll(entities.map((e) => new ds.Mutation()..update = e.toApiObject()));
+    return this;
+  }
 
   /// Adds multiple `upsert` mutations to this batch.
   ///
@@ -70,37 +76,51 @@ class MutationBatch {
   /// is complete this is done regardless of whether an entity with the same key
   /// exists. The key can also be incomplete in which case a unique id will be
   /// generated and an insertion will be performed.
-  void upsertAll(Iterable<Entity> entities) => mutations.addAll(
-        entities.map((e) => new ds.Mutation()..upsert = e.toApiObject()),
-      );
+  MutationBatch upsertAll(Iterable<Entity> entities) {
+    mutations.addAll(entities.map((e) => new ds.Mutation()..upsert = e.toApiObject()));
+    return this;
+  }
 
   /// Adds multiple `delete` mutations to this batch.
   ///
   /// Deletes the entity with the key. The key must be complete. The entity with
   /// the key may or may not exist.
-  void deleteAll(Iterable<Key> keys) => mutations.addAll(
-        keys.map((key) => new ds.Mutation()..delete = key.toApiObject()),
-      );
+  MutationBatch deleteAll(Iterable<Key> keys) {
+    mutations.addAll(keys.map((key) => new ds.Mutation()..delete = key.toApiObject()));
+    return this;
+  }
 
-  /// Executes all the mutations.
-  Future<MutationBatchResponse> execute() => executeRaw().then(
-        (ds.CommitResponse resp) => new MutationBatchResponse(shell, resp));
+  /// Commits all the mutations.
+  ///
+  /// Throws [ConcurrentModificationError] if the transaction has failed.
+  Future<MutationBatchResponse> commit() => commitRaw().then((ds.CommitResponse resp) {
+        if (resp.mutationResults.any((ds.MutationResult result) => result.conflictDetected))
+          throw new ConcurrentModificationError();
+        return new MutationBatchResponse(shell, resp);
+      });
+
+  @deprecated
+  Future<MutationBatchResponse> execute() => commit();
 
   /// Creates a `package:googleapis` request object for this batch.
   ///
-  /// Normally you would need [execute], this is for tinkering with the request
+  /// Normally you would need [commit], this is for tinkering with the request
   /// before sending it out.
   ds.CommitRequest createRequest() => new ds.CommitRequest()
     ..mutations = mutations
-    ..mode = "NON_TRANSACTIONAL";
+    ..transaction = shell.transactionId
+    ..mode = shell.isTransactional ? "TRANSACTIONAL" : "NON_TRANSACTIONAL";
 
-  /// Executes all the mutations and returns asynchronically with the raw `package:googleapis`
+  /// Commits all the mutations and returns asynchronically with the raw `package:googleapis`
   /// response object.
-  Future<ds.CommitResponse> executeRaw() {
-        DatastoreShell shell = this.shell;
+  Future<ds.CommitResponse> commitRaw() {
+    DatastoreShell shell = this.shell;
     ds.CommitRequest req = createRequest();
     return shell.api.projects.commit(req, shell.project);
   }
+
+  @deprecated
+  Future<ds.CommitResponse> executeRaw() => commitRaw();
 
   /// Provides an access to all the mutations the object has generated for tinkering.
   List<ds.Mutation> mutations = [];
@@ -112,8 +132,7 @@ class MutationBatch {
 /// The result of a mutation is an array of keys which are set to either the generated keys
 /// or to `null` if no key generation was necessary.
 class MutationBatchResponse {
-  MutationBatchResponse(this.shell, ds.CommitResponse response)
-      : _size = response.mutationResults.length {
+  MutationBatchResponse(this.shell, ds.CommitResponse response) : _size = response.mutationResults.length {
     int index = -1;
     _hasConflicts = false;
     response.mutationResults.forEach((ds.MutationResult result) {